+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.9-21] - 2020-07-10
 ### Added
 - StoryConfig: caution, sites info
 - StoryCommand: motif, foreshadow, payoff
 - StoryConfig: copy, theme and more
+- SceneInfoUpdater
 ### Changed
 - Compiler: conv to plot (re-format)
 ### Fixed
@@ -592,7 +595,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example story as usage.
 - Output story as markdown.
 
-[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-20...HEAD
+[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.5.9-21...HEAD
+[0.5.9-21]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-21
 [0.5.9-20]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-20
 [0.5.9-19]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-19
 [0.5.9-18]: https://github.com/nagisc007/storybuilder/releases/v0.5.9-18